DOC: Fixed dtype docs for var, nanvar.
Docs incorrectly stated that the default output dtype was float32 when
input array is of an integer dtype. However:

```python
np.var(np.random.randint(0, 100, 1000)).dtype == np.dtype("f8")
```

+++ numpy/core/fromnumeric.py
@@ -3419,7 +3419,7 @@ def var(a, axis=None, dtype=None, out=None, ddof=0, keepdims=np._NoValue):
         instead of a single axis or all the axes as before.
     dtype : data-type, optional
         Type to use in computing the variance.  For arrays of integer type
-        the default is `float32`; for arrays of float types it is the same as
+        the default is `float64`; for arrays of float types it is the same as
         the array type.
     out : ndarray, optional
         Alternate output array in which to place the result.  It must have
